requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langgraph
duckduckgo-search
python-dotenv
pandas
sqlalchemy
amadeus
requests
wikipedia
google-auth-oauthlib
google-auth
phonenumbers
bcrypt
streamlit-autorefresh
